+++ src/WalkingTec.Mvvm.Core/BaseCRUDVM.cs
@@ -1036,12 +1036,15 @@ namespace WalkingTec.Mvvm.Core
                 if (typeof(IWorkflow).IsAssignableFrom(typeof(TModel)))
                 {
                     var wi = DC.Set<Elsa_WorkflowInstance>().CheckEqual(typeof(TModel).FullName, x => x.ContextType).CheckEqual(Entity.GetID().ToString(), x => x.ContextId).ToList();
-                    DC.Set<Elsa_WorkflowInstance>().RemoveRange(wi);
-                    var wl = DC.Set<Elsa_WorkflowExecutionLogRecord>().CheckContain(wi.Select(x=>x.ID).ToList(), x => x.WorkflowInstanceId).ToList();
-                    DC.Set<Elsa_WorkflowExecutionLogRecord>().RemoveRange(wl);
-                    var ww = DC.Set<FrameworkWorkflow>().CheckContain(wi.Select(x => x.ID).ToList(), x => x.WorkflowId).ToList();
-                    DC.Set<FrameworkWorkflow>().RemoveRange(ww);
-                    DC.SaveChanges();
+                    if (wi.Count > 0)
+                    {
+                        DC.Set<Elsa_WorkflowInstance>().RemoveRange(wi);
+                        var wl = DC.Set<Elsa_WorkflowExecutionLogRecord>().CheckContain(wi.Select(x => x.ID).ToList(), x => x.WorkflowInstanceId).ToList();
+                        DC.Set<Elsa_WorkflowExecutionLogRecord>().RemoveRange(wl);
+                        var ww = DC.Set<FrameworkWorkflow>().CheckContain(wi.Select(x => x.ID).ToList(), x => x.WorkflowId).ToList();
+                        DC.Set<FrameworkWorkflow>().RemoveRange(ww);
+                        DC.SaveChanges();
+                    }
                 }
                 if (Wtm.ServiceProvider != null)
                 {
@@ -1099,6 +1102,19 @@ namespace WalkingTec.Mvvm.Core
                 }
                 DC.DeleteEntity(Entity);
                 await DC.SaveChangesAsync();
+                if (typeof(IWorkflow).IsAssignableFrom(typeof(TModel)))
+                {
+                    var wi = DC.Set<Elsa_WorkflowInstance>().CheckEqual(typeof(TModel).FullName, x => x.ContextType).CheckEqual(Entity.GetID().ToString(), x => x.ContextId).ToList();
+                    if (wi.Count > 0)
+                    {
+                        DC.Set<Elsa_WorkflowInstance>().RemoveRange(wi);
+                        var wl = DC.Set<Elsa_WorkflowExecutionLogRecord>().CheckContain(wi.Select(x => x.ID).ToList(), x => x.WorkflowInstanceId).ToList();
+                        DC.Set<Elsa_WorkflowExecutionLogRecord>().RemoveRange(wl);
+                        var ww = DC.Set<FrameworkWorkflow>().CheckContain(wi.Select(x => x.ID).ToList(), x => x.WorkflowId).ToList();
+                        DC.Set<FrameworkWorkflow>().RemoveRange(ww);
+                        await DC.SaveChangesAsync();
+                    }
+                }
                 if (Wtm.ServiceProvider != null)
                 {
                     var fp = Wtm.ServiceProvider.GetRequiredService<WtmFileProvider>();
